requirements.txt
accelerate==1.7.0
bitsandbytes==0.43.1
deepspeed==0.14.4
datasets==2.20.0
einops==0.8.1
fire==0.7.0
huggingface-hub==0.31.4
matplotlib==3.10.3
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.2.3
peft==0.11.1
rich==14.0.0
safetensors==0.4.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
tokenizers==0.21.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.49.0
triton==3.0.0
wandb==0.19.11
pillow==10.4.0
prettytable
ipython
timm
tabulate
